Remove dep on vault internals
Vault dependency was clashing with pre-go mod version of Vault
used by the Terraform dep. The Terraform dep is used for hcl2shim
but is not meant to be imported and so imports the whole project
