+++ e2e/command_test.go
@@ -11,8 +11,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/hashicorp/consul-terraform-sync/api"
 	"github.com/hashicorp/consul-terraform-sync/testutils"
-	ctsTestClient "github.com/hashicorp/consul-terraform-sync/testutils/cts"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -34,7 +34,7 @@ func TestE2E_MetaCommandErrors(t *testing.T) {
 	config := fakeHandlerConfig().appendConsulBlock(srv).appendTerraformBlock(tempDir)
 	config.write(t, configPath)
 
-	cts, stop := ctsTestClient.StartCTS(t, configPath, ctsTestClient.CTSDevModeFlag)
+	cts, stop := api.StartCTS(t, configPath, api.CTSDevModeFlag)
 	defer stop(t)
 	err := cts.WaitForAPI(5 * time.Second)
 	require.NoError(t, err)
@@ -99,7 +99,7 @@ func TestE2E_EnableTaskCommand(t *testing.T) {
 	config := disabledTaskConfig().appendConsulBlock(srv).appendTerraformBlock(tempDir)
 	config.write(t, configPath)
 
-	cts, stop := ctsTestClient.StartCTS(t, configPath, ctsTestClient.CTSDevModeFlag)
+	cts, stop := api.StartCTS(t, configPath, api.CTSDevModeFlag)
 	defer stop(t)
 	err := cts.WaitForAPI(5 * time.Second)
 	require.NoError(t, err)
@@ -161,7 +161,7 @@ func TestE2E_DisableTaskCommand(t *testing.T) {
 		appendConsulBlock(srv).appendDBTask()
 	config.write(t, configPath)
 
-	cts, stop := ctsTestClient.StartCTS(t, configPath, ctsTestClient.CTSDevModeFlag)
+	cts, stop := api.StartCTS(t, configPath, api.CTSDevModeFlag)
 	defer stop(t)
 	err := cts.WaitForAPI(10 * time.Second)
 	require.NoError(t, err)
