+++ e2e/e2e_test.go
@@ -9,7 +9,6 @@ import (
 	"time"
 
 	"github.com/hashicorp/consul-terraform-sync/testutils"
-	ctsTestClient "github.com/hashicorp/consul-terraform-sync/testutils/cts"
 	"github.com/hashicorp/consul/sdk/testutil"
 	"github.com/stretchr/testify/require"
 )
@@ -115,7 +114,7 @@ func TestE2ERestartConsul(t *testing.T) {
 	config.write(t, configPath)
 
 	// start CTS
-	cts, stop := ctsTestClient.StartCTS(t, configPath)
+	cts, stop := api.StartCTS(t, configPath)
 	defer stop(t)
 	// wait enough for cts to cycle through once-mode successfully
 	err := cts.WaitForAPI(15 * time.Second)
@@ -167,7 +166,7 @@ func TestE2EPanosHandlerError(t *testing.T) {
 		appendTerraformBlock(tempDir, requiredProviders)
 	config.write(t, configPath)
 
-	ctsTestClient.StartCTS(t, configPath, ctsTestClient.CTSOnceModeFlag)
+	api.StartCTS(t, configPath, api.CTSOnceModeFlag)
 
 	delete()
 }
@@ -247,7 +246,7 @@ func TestE2ELocalBackend(t *testing.T) {
 			configPath := filepath.Join(tempDir, configFile)
 			config.write(t, configPath)
 
-			ctsTestClient.StartCTS(t, configPath, ctsTestClient.CTSOnceModeFlag)
+			api.StartCTS(t, configPath, api.CTSOnceModeFlag)
 
 			// check that statefile was created locally
 			checkStateFileLocally(t, tc.dbStateFilePath)
@@ -274,7 +273,7 @@ func newTestConsulServer(t *testing.T) *testutil.TestServer {
 }
 
 func runSyncStop(t *testing.T, configPath string, dur time.Duration) {
-	cts, stop := ctsTestClient.StartCTS(t, configPath)
+	cts, stop := api.StartCTS(t, configPath)
 	cts.WaitForAPI(dur)
 	stop(t)
 }
